Module Expect_test_collector

module Test_outcome : sig ... end

module Make (Config : Expect_test_config_types.S) : sig ... end

val am_running_expect_test : unit -> bool

Returns true if and only if an expect test is currently collecting output.

val save_and_return_output : Expect_test_common.File.Location.t -> string

Flushes stdout/stderr. Same as Make().save_and_return_output, without monad.

val tests_run : unit -> Test_outcome.t list

The tests that ran, in the order they ran

module Current_file : sig ... end