Parameter Make.Config

module IO : sig ... end

val run : (unit -> unit IO.t) -> unit

Run an IO operation until completion

val sanitize : string -> string

sanitize can be used to map all output strings, e.g. for cleansing.

val upon_unreleasable_issue : [ `CR | `Warning_for_collector_testing ]

upon_unreleasable_issue specifies how to deal with output that should not be released even if it is accepted (e.g. backtraces). The default is `CR.